requirements.txt
Django==5.0.1
Pillow>=10.4.0
reportlab==4.0.7
openpyxl==3.1.2
django-environ==0.11.2
djangorestframework==3.14.0
django-redis==5.4.0
drf-yasg==1.21.7


